requirements.txt
click
python-dotenv
google-generativeai